# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/issue-type": {
      "post": {
        "operationId": "IssueTypeController_create",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Get issue type list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTypeDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "IssueType"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "IssueTypeController_findAll",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get issue type list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssueTypeDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "IssueType"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/issue-type/{id}": {
      "get": {
        "operationId": "IssueTypeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get issue type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTypeDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "IssueType"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "IssueTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssueTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get issue type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueTypeDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "IssueType"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "IssueTypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete issue type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "IssueType"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Manager API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DeleteDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "CreateIssueTypeColumnDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "filedName": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          },
          "issueTypeId": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "filedName",
          "isRequired",
          "issueTypeId"
        ]
      },
      "CreateIssueTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "Bug",
              "Task",
              "Sub-task",
              "Spike",
              "Test",
              "Research"
            ],
            "default": "Task"
          },
          "issueTypeColumns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateIssueTypeColumnDto"
            }
          }
        },
        "required": [
          "name",
          "description",
          "icon",
          "color",
          "isActive",
          "type"
        ]
      },
      "IssueTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "Bug",
              "Task",
              "Sub-task",
              "Spike",
              "Test",
              "Research"
            ],
            "default": "Task"
          },
          "issueTypeColumns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueTypeColumnDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "icon",
          "color",
          "isActive",
          "type",
          "issueTypeColumns",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "IssueTypeColumnDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "filedName": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          },
          "issueTypeId": {
            "type": "number"
          },
          "issueType": {
            "$ref": "#/components/schemas/IssueTypeDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "filedName",
          "type",
          "isRequired",
          "issueTypeId",
          "issueType",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateIssueTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "Bug",
              "Task",
              "Sub-task",
              "Spike",
              "Test",
              "Research"
            ],
            "default": "Task"
          },
          "issueTypeColumns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateIssueTypeColumnDto"
            }
          }
        }
      },
...
    }
  }
}